Fixes for the race conditions in the chart
Chart can now be installed and deleted
Chart keeps the lvms namespace because we don't want to remove storage
Known error with the forklift-api which is missing while running the web
hook might say the installation fail. We need to ignore this right now
because the installation eventually succeeds and mtv is running.

Signed-off-by: Roy Golan <rgolan@redhat.com>